### 2021-09-16 #463 @dieghernan
Submission: {cffr} Generate Citation File Format ('cff') Metadata for R Packages
Submitting Author Name: Diego Hernangómez
Submitting Author Github Handle: @dieghernan
Repository: https://github.com/dieghernan/cffr
Version submitted: 0.0.1.9001
Submission type: Standard
Version accepted: TBD

---

```
Package: cffr
Title: Generate Citation File Format ('cff') Metadata for R Packages
Version: 0.0.1.9001
    person(given = "Diego",
           family = "Hernangómez",
           email = "diego.hernangomezherrero@gmail.com",
Description: The Citation File Format version 1.2.0 (Druskat, S. et al.
    (2021) <doi:10.5281/zenodo.5171937>) is a human and machine readable
    file format which provides citation metadata for software. This
    package provides core utilities to generate and validate this
    metadata.
    https://github.com/dieghernan/cffr
    crayon (>= 1.4.1),
    desc (>= 1.3.0),
    jsonlite (>= 1.7.2),
    jsonvalidate (>= 1.1.0),
    yaml (>= 2.2.1)
    knitr
X-schema.org-keywords: attribution, citation, credit, citation-files, cff,
    metadata
```

The package extract metadata from DESCRIPTION and CITATION and generates a [CITATION.cff file](https://citation-file-format.github.io/).

CITATION.cff files are plain text files with human- and machine-readable citation information for software (and datasets).

Any R package developer. A CITATION.cff file provides clarity when citating R packages. [GitHub](https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/creating-a-repository-on-github/about-citation-files), [Zenodo](https://twitter.com/ZENODO_ORG/status/1420357001490706442) and [Zotero](https://twitter.com/zotero/status/1420515377390530560) supports this format.

[**{citation}**](https://github.com/pik-piam/citation/): The development version (at the time of this writing) includes a new function `r2cff` that creates a minimal `CITATION.cff` file (v1.1.0) using the information of your  `DESCRIPTION` file. It also provide minimal validity checks.

[**{handlr}**](https://github.com/ropensci/handlr): Tool for converting among citation formats, including `*.cff` files. At the time of
  this writing only CFF v1.1.0 was supported (see [#24](https://github.com/ropensci/handlr/issues/24)).

**{cffr}** performs an extensive extraction of DESCRIPTION and CITATION, as well as validation via **{jsonvalidate}** against the
official [schema.json](https://github.com/citation-file-format/citation-file-format/blob/main/schema.json). None of the packages mentioned before supports CFF v1.2.0, that is the latest released version.

It also provides parsers for `person` and `citation/bibentry` classes.

N/A

https://github.com/ropensci/software-review/issues/462 @noamross
