+++ .github/workflows/test-ci.yaml
@@ -49,13 +49,6 @@ jobs:
           packs <- dplyr::bind_rows(packs)
           core <- dplyr::filter(packs, core == TRUE)
 
-          # rOpenSci packages from r-universe
-
-          options(repos = c(
-            ropensci = "https://ropensci.r-universe.dev",
-            CRAN = "https://cloud.r-project.org"
-          ))
-
           ropensci <- unlist(jsonlite::read_json("https://ropensci.r-universe.dev/packages"))
 
           all <- sort(unique(c(core$name, ropensci)))
