+++ .github/workflows/test-ci.yaml
@@ -8,7 +8,7 @@ name: Test on GHA
 
 jobs:
   test-ci:
-    runs-on: macos-latest
+    runs-on: linux-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
@@ -29,6 +29,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
+          cache-version: 900
           extra-packages: |
             devtools
             remotes
