+++ .github/workflows/test-ci.yaml
@@ -8,7 +8,7 @@ name: Test on GHA
 
 jobs:
   test-ci:
-    runs-on: linux-latest
+    runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
