+++ .github/workflows/test-ci.yaml
@@ -8,7 +8,7 @@ name: Test on GHA
 
 jobs:
   test-ci:
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
@@ -19,6 +19,7 @@ jobs:
       - uses: r-lib/actions/setup-r@v1
         with:
           use-public-rspm: true
+          install-r: false
 
       # This is needed for workflows running on
       # ubuntu-20.04 or later
@@ -67,7 +68,7 @@ jobs:
           toinstall <- setdiff(all, instpack)
           
           install.packages(toinstall, dependencies = TRUE, 
-                           verbose = TRUE, quiet = TRUE)
+                           verbose = TRUE, quiet = TRUE, type="binary")
           
           
         shell: Rscript {0}
