### 2021-09-17 #463 @emilyriederer
Hi @dieghernan - so nice to meet you and thank you for submitting this great package!

I did not notice any causes for concern on my first review, and the only enhancement found by the testing suite is perhaps splitting up a few of the long lines of code found by [`lintr`](https://github.com/jimhester/lintr). You can run `lintr::lint_package(".")`  within your R package to reproduce these locally.

I think we are ready to move to the next stage, so I will begin to look for reviewers. We will be in touch soon!
