### 2021-09-17 7169202
Avoiding long lines of code
See https://github.com/ropensci/software-review/issues/463

#4
