+++ .github/workflows/lint.yaml
@@ -19,7 +19,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v1
         with:
           use-public-rspm: true
-          install-r: false
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
