+++ .github/workflows/pkgdown-gh-pages-clean.yaml
@@ -16,7 +16,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v1
         with:
           use-public-rspm: true
-          install-r: false
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
