+++ .github/workflows/update-citation-cff.yaml
@@ -5,7 +5,7 @@ on:
       - DESCRIPTION
       - inst/CITATION
   workflow_dispatch:
-  
+
 name: Update CITATION.cff
 
 jobs:
@@ -19,7 +19,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v1
         with:
           use-public-rspm: true
-          install-r: false
 
       # This is needed for workflows running on
       # ubuntu-20.04 or later
@@ -36,16 +35,16 @@ jobs:
 # When on CRAN, use instead
          # extra-packages: |
          #  cffr
-     
+
 
       - name: Update CITATION.cff
         run: |
 
           library(cffr)
-          
+
           # Customize with your own code
           # Mind the indentation!
-          
+
           cff_write()
 
         shell: Rscript {0}
