+++ .github/workflows/pkgdown-gh-pages.yaml
@@ -20,7 +20,6 @@ jobs:
       - uses: r-lib/actions/setup-r@v1
         with:
           use-public-rspm: true
-          install-r: false
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
