+++ README.md
@@ -16,6 +16,7 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 bytes](https://img.shields.io/github/languages/code-size/dieghernan/cffr)
 [![CodeFactor](https://www.codefactor.io/repository/github/dieghernan/cffr/badge)](https://www.codefactor.io/repository/github/dieghernan/cffr)
 
+[![peer-review](https://badges.ropensci.org/463_status.svg)](https://github.com/ropensci/software-review/issues/463)
 <!-- badges: end -->
 
 [Citation File Format (CFF)](https://citation-file-format.github.io/)
