### 2021-09-17 #463 @dieghernan
Thanks @emilyriederer , nice to meet you as well.

I did my best trying to remove long lines of code, most of them were links to long urls, but I managed to get no lint warning on that, see:

https://github.com/dieghernan/cffr/actions/runs/1244840358

I also added the badge to the README, thanks for the `rodev::use_review_badge(463)`

(PS: I was looking for images of neural networks and I found that fantastic .png 😄 )
