+++ .github/workflows/test-ci.yaml
@@ -55,7 +55,8 @@ jobs:
           })
 
           packs <- dplyr::bind_rows(packs)
-          core <- dplyr::filter(packs, core == TRUE)
+          #core <- dplyr::filter(packs, core == TRUE)
+          core <- packs
 
           # rOpenSci packages from r-universe
 
