### 2021-09-17 98d4125
Vectorise parsers
