+++ R/cff_create.R
@@ -152,7 +152,7 @@ cff_create <- function(x = ".", keys = NULL,
 
     cffobj <- cff_description(desc_path, cff_version)
   }
-  
+
   citobj <- unique(citobj)
   # Add cffobj
 
