### 2021-09-17 241908a
Close current cycle of development. Stable
