+++ tests/testthat/_snaps/cff_parse_person.md
@@ -1,0 +1,14 @@
+# Parse one person
+
+    family-names: person
+    given-names: one
+
+# Parse several persons
+
+    - family-names: person
+      given-names: one
+    - family-names: human
+      given-names: another
+    - family-names: more
+      given-names: and one
+
