### 2021-09-17 2499c03
Fix bug with citations
