+++ R/cff_parse_citation.R
@@ -160,10 +160,8 @@ cff_parse_citation <- function(bib) {
 
   # Keywords
   if ("keywords" %in% names(parse_cit)) {
-    newkeys <- unique(unlist(strsplit(parse_cit$keywords, ", ")))
-    newkeys2 <- unique(unlist(strsplit(parse_cit$keywords, ",")))
-    newkeysend <- unique(newkeys, newkeys2)
-    parse_cit$keywords <- newkeysend
+    newkeys <- trimws(unique(unlist(strsplit(parse_cit$keywords, ","))))
+    parse_cit$keywords <- newkeys
   }
 
   #  Other persons
