### 2021-09-17 #463 @emilyriederer
Thanks, @dieghernan !

One other metric I just wanted to call out is the high cyclocomplexity scores for `clean_str()` and `cff_create()`.

If you aren't familiar with this score (I wasn't until this review!), the R package's [README](https://github.com/MangoTheCat/cyclocomp) gives some good examples. In short, it measure how many different distinct paths your code might go through when evaluation (based on things like `if-else` clauses). Right now, our checks highlight scores greater than 15. High scores might suggest opportunities for simplification with strategies like using more internal helper functions; high scores can also suggest that the function is hard to test robustly since it can be hard to hit all the corner cases.

I want to say upfront, **I am comfortable with your code as-is and think these functions are very reasonable**. For your functions, all of the blocks seem necessary and so concise that more internal functions might actually decrease the readability of the code.

That said, I thought I'd also highlight that part of the review for your judgement.

Still working on the reviewers but hope to have some news soon!
