### 2021-09-17 #463 @dieghernan
Thanks @emilyriederer, I would have a look on that. I still don’t fully understand the high score on `clean_str` (24 loc with 3 if/else, would investigate it).

I would like to highlight that I have implemented a strategy with `testthat` + GHA for testing the package on >5000 packages. More information here:
- https://github.com/dieghernan/cffr/tree/main/tests/testthat/test_ci

Latest results here:

- https://github.com/dieghernan/cffr/actions/runs/1247163030

I have used this for detecting  corner cases, but probably this have added also complexity on the functions.

The tests are OK now, and in the packages that the validation is failing are due to errors on the raw files of the packages tested (i.e. malformed urls and DOI/ORCID on old packages, that won’t be compliant with the current CRAN policies).

Thanks for your hard work. May I suggest for review any of the authors of {codemeta} or {handlr} (both in rOpenSci)? The scope of these packages are closely related with {cffr} and it will be interesting to have their feedback.
