+++ dev/testlocal.R
@@ -7,10 +7,9 @@ setwd("~/R/Projects/cffr")
 devtools::load_all()
 
 # Delete first the snaps folder
-unlink("./tests/testthat/test_local_installation/_snaps",
+unlink("./tests/testthat/test_ci/_snaps",
        recursive = TRUE
 )
 
 # Run the tests
-testthat::test_dir("tests/testthat/test_local_installation")
-
+testthat::test_dir("tests/testthat/test_ci")
