### 2021-09-18 #463 @dieghernan
I @emilyriederer , I have introduced some changes and removing some unneeded chunks (https://github.com/dieghernan/cffr/commit/1aeb1868b8c3989fe8e9e9b438a7d99658270034)

Now running `cyclocomp::cyclocomp_package_dir()` I see:

|   |name                          | cyclocomp|
|:--|:-----------------------------|---------:|
|7  |cff_create                    |        13|
|17 |cff_write                     |        11|
|9  |cff_parse_citation            |        10|
|10 |cff_parse_person              |         8|
|19 |clean_str                     |         8|
|6  |cff                           |         7|
|35 |parse_desc_urls               |         6|
|42 |rapply_cff                    |         6|
|1  |as.cff                        |         5|
|16 |cff_validate                  |         5|
|39 |parse_r_citation              |         5|
|23 |is.substring                  |         4|
|3  |building_month                |         3|
|5  |building_url                  |         3|
|18 |clean_jsonlite                |         3|
|22 |is.email                      |         3|
|24 |is.url                        |         3|
|25 |merge_desc_cit                |         3|
|32 |parse_desc_keywords           |         3|
|2  |building_doi                  |         2|
|14 |cff_schema_keys               |         2|
|21 |is.cff                        |         2|
|29 |parse_desc_authors            |         2|
|31 |parse_desc_date_released      |         2|
|37 |parse_package_meta            |         2|
|38 |parse_preferred_auto          |         2|
|43 |stopifnotcff                  |         2|
|44 |stopifnotexists               |         2|
|4  |building_other_persons        |         1|
|8  |cff_description               |         1|
|11 |cff_schema_definitions_entity |         1|
|12 |cff_schema_definitions_person |         1|
|13 |cff_schema_definitions_refs   |         1|
|15 |cff_schema_keys_license       |         1|
|20 |drop_null                     |         1|
|26 |other_persons                 |         1|
|27 |other_persons_entity          |         1|
|28 |parse_desc_abstract           |         1|
|30 |parse_desc_contacts           |         1|
|33 |parse_desc_license            |         1|
|34 |parse_desc_title              |         1|
|36 |parse_desc_version            |         1|
|40 |print.cff                     |         1|
|41 |print_snapshot                |         1|
|45 |validate_schema               |         1|
