+++ tests/testthat/_snaps/parse_citation.md
@@ -21,24 +21,18 @@
         doi = {10.1111/2041-210X.12469},
       }
     
-    To cite RNeXML in publications, please use:
-    
-    Boettiger C, Chamberlain S, Vos R, Lapp H (2016). "RNeXML: A Package
-    for Reading and Writing Richly Annotated Phylogenetic, Character, and
-    Trait Data in R." _Methods in Ecology and Evolution_, *7*, 352-357.
-    doi: 10.1111/2041-210X.12469 (URL:
-    https://doi.org/10.1111/2041-210X.12469).
+      H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
+      Springer-Verlag New York, 2016.
     
     A BibTeX entry for LaTeX users is
     
-      @Article{,
-        title = {{RNeXML}: {A} Package for Reading and Writing Richly Annotated Phylogenetic, Character, and Trait Data in {R}},
-        journal = {Methods in Ecology and Evolution},
-        author = {Carl Boettiger and Scott Chamberlain and Rutger Vos and Hilmar Lapp},
+      @Book{,
+        author = {Hadley Wickham},
+        title = {ggplot2: Elegant Graphics for Data Analysis},
+        publisher = {Springer-Verlag New York},
         year = {2016},
-        volume = {7},
-        pages = {352--357},
-        doi = {https://doi.org/10.1111/2041-210X.12469},
+        isbn = {978-3-319-24277-4},
+        url = {https://ggplot2.tidyverse.org},
       }
     
 
@@ -65,24 +59,18 @@
         doi = {10.1111/2041-210X.12469},
       }
     
-    To cite RNeXML in publications, please use:
-    
-    Boettiger C, Chamberlain S, Vos R, Lapp H (2016). "RNeXML: A Package
-    for Reading and Writing Richly Annotated Phylogenetic, Character, and
-    Trait Data in R." _Methods in Ecology and Evolution_, *7*, 352-357.
-    doi: 10.1111/2041-210X.12469 (URL:
-    https://doi.org/10.1111/2041-210X.12469).
+      H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
+      Springer-Verlag New York, 2016.
     
     A BibTeX entry for LaTeX users is
     
-      @Article{,
-        title = {{RNeXML}: {A} Package for Reading and Writing Richly Annotated Phylogenetic, Character, and Trait Data in {R}},
-        journal = {Methods in Ecology and Evolution},
-        author = {Carl Boettiger and Scott Chamberlain and Rutger Vos and Hilmar Lapp},
+      @Book{,
+        author = {Hadley Wickham},
+        title = {ggplot2: Elegant Graphics for Data Analysis},
+        publisher = {Springer-Verlag New York},
         year = {2016},
-        volume = {7},
-        pages = {352--357},
-        doi = {https://doi.org/10.1111/2041-210X.12469},
+        isbn = {978-3-319-24277-4},
+        url = {https://ggplot2.tidyverse.org},
       }
     
 
