+++ tests/testthat/_snaps/merge_desc_cit.md
@@ -1,0 +1,734 @@
+# Merge all DESCRIPTION files with CITATION_basic
+
+    cff-version: 1.2.0
+    message: 'To cite package "basicdesc" in publications use:'
+    type: software
+    title: 'basicdesc: A Basic Description'
+    version: 0.1.6
+    authors:
+    - family-names: Basic
+      given-names: Marc
+      email: marcbasic@gmail.com
+    abstract: A very basic description. Should parse without problems.
+    repository-code: https://github.com/basic/package
+    url: https://basic.github.io/package
+    contact:
+    - family-names: Basic
+      given-names: Marc
+      email: marcbasic@gmail.com
+    license: GPL-3.0-only
+    doi: 10.1111/2041-210X.12469
+    preferred-citation:
+      type: article
+      title: 'RNeXML: A Package for Reading and Writing Richly Annotated Phylogenetic,
+        Character, and Trait Data in R'
+      authors:
+      - family-names: Boettiger
+        given-names: Carl
+      - family-names: Chamberlain
+        given-names: Scott
+      - family-names: Vos
+        given-names: Rutger
+      - family-names: Lapp
+        given-names: Hilmar
+      journal: Methods in Ecology and Evolution
+      year: '2016'
+      volume: '7'
+      doi: 10.1111/2041-210X.12469
+    references:
+    - type: book
+      title: 'ggplot2: Elegant Graphics for Data Analysis'
+      authors:
+      - family-names: Wickham
+        given-names: Hadley
+      publisher:
+        name: Springer-Verlag New York
+      year: '2016'
+      isbn: 978-3-319-24277-4
+      url: https://ggplot2.tidyverse.org
+
+---
+
+    cff-version: 1.2.0
+    message: 'To cite package "basicdesc" in publications use:'
+    type: software
+    title: 'basicdesc: A Basic Description'
+    version: 0.1.6
+    authors:
+    - family-names: Basic
+      given-names: Marc
+      email: marcbasic@gmail.com
+    abstract: A very basic description. Should parse without problems.
+    repository-code: https://github.com/basic/package
+    url: https://basic.github.io/package
+    contact:
+    - family-names: Basic
+      given-names: Marc
+      email: marcbasic@gmail.com
+    license: GPL-3.0-only
+    doi: 10.1111/2041-210X.12469
+    preferred-citation:
+      type: article
+      title: 'RNeXML: A Package for Reading and Writing Richly Annotated Phylogenetic,
+        Character, and Trait Data in R'
+      authors:
+      - family-names: Boettiger
+        given-names: Carl
+      - family-names: Chamberlain
+        given-names: Scott
+      - family-names: Vos
+        given-names: Rutger
+      - family-names: Lapp
+        given-names: Hilmar
+      journal: Methods in Ecology and Evolution
+      year: '2016'
+      volume: '7'
+      doi: 10.1111/2041-210X.12469
+    references:
+    - type: book
+      title: 'ggplot2: Elegant Graphics for Data Analysis'
+      authors:
+      - family-names: Wickham
+        given-names: Hadley
+      publisher:
+        name: Springer-Verlag New York
+      year: '2016'
+      isbn: 978-3-319-24277-4
+      url: https://ggplot2.tidyverse.org
+
+---
+
+    cff-version: 1.2.0
+    message: 'To cite package "basicdescdate" in publications use:'
+    type: software
+    title: 'basicdescdate: A Basic Description with Date'
+    version: 0.1.6
+    authors:
+    - family-names: Basic
+      given-names: Marc
+      email: marcbasic@gmail.com
+    abstract: A very basic description. Should parse without problems. I have a Date
[…]
