+++ man/cff_create.Rd
@@ -84,6 +84,8 @@ cff_create(demo_file, keys = list("contact" = new_contact))
 \seealso{
 \href{https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md}{Guide to Citation File Format schema version 1.2.0}.
 
+\code{vignette("cffr", "cffr")}
+
 Other core functions: 
 \code{\link{cff_validate}()},
 \code{\link{cff_write}()},
