+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(c,cff)
 S3method(print,cff)
 export(as.cff)
 export(cff)
