+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: cffr
 Title: Generate Citation File Format ('cff') Metadata for R Packages
-Version: 0.0.1.9001
+Version: 0.0.1.9002
 Authors@R: 
     person(given = "Diego",
            family = "Hernangómez",
