+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: cffr
 Title: Generate Citation File Format ('cff') Metadata for R Packages
-Version: 0.0.1.9002
+Version: 0.0.1.9003
 Authors@R: 
     person(given = "Diego",
            family = "Hernangómez",
