### 2021-09-21 a10498b
Improve repository detection
