+++ R/utils.R
@@ -45,3 +45,52 @@ print_snapshot <- function(title = "----", obj) {
   print(obj)
   cat("\n---")
 }
+
+#' Search a package on available repos
+#' @noRd
+search_on_repos <- function(name) {
+  repos <- detect_repos()
+
+
+  get <- avail[name == avail$Package, c("Repository")]
+
+  get <- clean_str(get)
+
+  if (is.null(get)) {
+    return(NULL)
+  }
+
+  # Try to find in CRAN
+  cran_repo <- clean_str(repos["CRAN"])
+
+
+  if (length(grep(cran_repo, get) == 1)) {
+    # Canonic url to CRAN
+
+    repos <- paste0("https://cran.r-project.org/package=", name)
+    return(repos)
+  }
+
+  repos <- gsub("src/contrib$", "", get)
+
+  repos
+}
+
+
+#' Detect current repos
+#' @noRd
+detect_repos <- function() {
+  repos <- getOption("repos")
+
+  # If not set use 0-Cloud
+  if (!is.substring(repos["CRAN"], "^http")) {
+    repos["CRAN"] <- "https://cloud.r-project.org/"
+  }
+  return(repos)
+}
+
+
+# Load this one per session
+avail <- as.data.frame(available.packages(
+  repos = detect_repos(),
+))[, c("Package", "Repository")]
