### 2021-09-21 67925d3
Fix canonic URLs of CRAN
