+++ vignettes/crosswalk.Rmd
@@ -410,7 +410,7 @@ print(citation("rmarkdown")[-1], bibtex = TRUE)
 
 This key is extracted from the "Repository" field of the DESCRIPTION file. 
 Usually, this field is auto-populated when a package is hosted on a repo (like
-CRAN or the [r-universe](https://r-universe.dev/). For packages without this
+CRAN or the [r-universe](https://r-universe.dev/)). For packages without this
 field on the DESCRIPTION (that is the typical case for an in-development
 package), **cffr** would try to search the package on any of the 
 default repositories specified on  `options("repos")`.
