+++ .github/workflows/update-docs.yaml
@@ -5,7 +5,7 @@ name: update-docs
 
 jobs:
   update-docs:
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
@@ -27,13 +27,9 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
           extra-packages: |
-            devtools
             rcmdcheck
             pkgdown
             dieghernan/pkgdev
-            tidyverse
-            scales
-            reactable
           needs: website
 
       - name: Update docs
