+++ .github/workflows/pkgdown-gh-pages.yaml
@@ -9,7 +9,7 @@ name: pkgdown-gh-pages
 
 jobs:
   pkgdown-gh-pages:
-    runs-on: ubuntu-latest
+    runs-on: macos-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
@@ -25,10 +25,6 @@ jobs:
         with:
           extra-packages: |
             pkgdown
-            devtools
-            reactable
-            tidyverse
-            scales
           needs: website
 
       - name: Deploy package
