+++ R/cff.R
@@ -33,7 +33,7 @@
 #'
 #' # Use custom params
 #' test <- cff(
-#'   message = "A blank file",
+#'   title = "Manipulating files",
 #'   keywords = c("A", "new", "list", "of", "keywords"),
 #'   authors = list(cff_parse_person("New author"))
 #' )
@@ -46,7 +46,7 @@
 #' # Modify with cff_create
 #' new <- cff_create(test, keys = list(
 #'   "cff-version" = "1.2.0",
-#'   title = "Manipulating files"
+#'   message = "A blank file"
 #' ))
 #' new
 #'
@@ -91,10 +91,17 @@ cff <- function(path, ...) {
 #'
 #' @examples
 #'
-#' as.cff(list(
+#'
+#' # Convert a list to "cff" object
+#' cffobj <- as.cff(list(
 #'   "cff-version" = "1.2.0",
 #'   title = "Manipulating files"
 #' ))
+#'
+#' class(cffobj)
+#'
+#' # Nice display thanks to yaml package
+#' cffobj
 as.cff <- function(x) {
   if (is.cff(x)) {
     return(x)
