+++ vignettes/crosswalk.Rmd
@@ -22,15 +22,17 @@ library(cffr)
 
 ```
 
-The goal of this vignette is to provide clarity on how **cffr** extracts the
-metadata of a package and parse it to a valid `CITATION.cff` file.
+The goal of this vignette is to provide an explicit map between the metadata
+fields used by **cffr** and each one of the valid keys of the 
+[Citation File Format schema version
+1.2.0](https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md#valid-keys).
 
 
 ## Summary
 
 we summarize here the fields that **cffr** can parse and the original source of
 information for each one of them. The details on each key are presented on the
-next section of the document. The assessment of fields are based om the [Guide
+next section of the document. The assessment of fields are based on the [Guide
 to Citation File Format schema version
 1.2.0](https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md#valid-keys) [@Druskat_Citation_File_Format_2021].
 
@@ -81,7 +83,7 @@ knitr::kable(df, escape = FALSE)
 
 This key is extracted from the "Description" field of the DESCRIPTION file.
 
-<details><summary>Example</summary>
+<details><summary><strong>Example</strong></summary>
 
 
 ```{r abstract}
@@ -103,12 +105,14 @@ cat(pkg$get("Description"))
 
 </details>
 
+[Back to summary](#summary).
+
 ### authors
 
 This key is parsed from the "Authors" or "Authors\@R" field of the DESCRIPTION
 file. Only persons with the role "aut" or "cre" are considered.
 
-<details><summary>Example</summary>
+<details><summary><strong>Example</strong></summary>
 
 ```{r authors}
 
@@ -126,12 +130,14 @@ authors[vapply(authors, function(x) {
 
 </details>
 
+[Back to summary](#summary).
+
 ### cff-version
 
 This value can be set via the parameters of the `cff_create()`/`cff_write()`
 functions:
 
-<details><summary>Example</summary>
+<details><summary><strong>Example</strong></summary>
 
 ```{r cffversion}
 
@@ -141,6 +147,8 @@ cat(cff_objv110$`cff-version`)
 ```
 
 </details>
+
+[Back to summary](#summary).
 
 ### commit
 
@@ -159,12 +167,15 @@ to CFF schema v1.2.0](https://github.com/citation-file-format/citation-file-form
 >
 
 
+[Back to summary](#summary).
+
+
 ### contact
 
 This key is parsed from the "Authors" or "Authors\@R" field of the DESCRIPTION
 file. Only persons with the role "cre" (i.e, the maintainer(s)) are considered.
 
-<details><summary>Example</summary>
+<details><summary><strong>Example</strong></summary>
 
 ```{r contact}
 
@@ -173,11 +184,12 @@ pkg <- desc::desc(file.path(find.package("rmarkdown"), "DESCRIPTION"))
 
 cff_obj$contact
 
-authors[vapply(authors, function(x) {
-  "cre" %in% x$role
-}, logical(1))]
+pkg$get_author()
 ```
 </details>
+
+
+[Back to summary](#summary).
 
 ### date-released
 
@@ -185,7 +197,7 @@ This key is parsed from the "Date" field or, if not present, from the
 "Date/Publication" field that is present on packages built on CRAN.
 
 
-<details><summary>Example</summary>
[…]
