+++ .github/workflows/pkgdown-gh-pages.yaml
@@ -9,7 +9,7 @@ name: pkgdown-gh-pages
 
 jobs:
   pkgdown-gh-pages:
-    runs-on: macos-latest
+    runs-on: ubuntu-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
     steps:
