+++ NEWS.md
@@ -1,12 +1,13 @@
 # cffr (development version)
 
 -   Add DOI <https://doi.org/10.5281/zenodo.5509766>
--   Citation of installed packages extracted using `citation()`
+-   Citation of installed packages extracted using `citation().`
 -   Auto-generating `preferred-citation` key from DESCRIPTION.
 -   Rename `cff_schema_definitions_reference()` to
     `cff_schema_definitions_refs()`.
 -   "repository" key is supported.
 -   Added vignette: `vignette("crosswalk", package = "cffr")`.
+-   Add support to Bioconductor packages.
 
 # cffr 0.0.1
 
