+++ .github/workflows/check-full.yaml
@@ -28,7 +28,7 @@ jobs:
           - {os: windows-latest, r: 'oldrel'}
           #- {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
-          - {os: macOS-latest, r: 'oldrel'}
+          #- {os: macOS-latest, r: 'oldrel'}
           - {os: ubuntu-20.04, r: 'release'}
           - {os: ubuntu-20.04, r: 'devel'}
           - {os: ubuntu-20.04, r: 'oldrel'}
