### 2021-09-28 #463 @dieghernan
Hi @sckott , nice to meet you. I just wanted to say hello and let you know that some days ago I added a new vignette "Crosswalk" that provides tecnical details and examples on how {cffr} extracts the metadata. I think this may be useful:

https://dieghernan.github.io/cffr/articles/crosswalk.html

Regards
