+++ .github/workflows/check-full.yaml
@@ -29,10 +29,11 @@ jobs:
           #- {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
           #- {os: macOS-latest, r: 'oldrel'}
-          - {os: ubuntu-20.04, r: 'release'}
-          - {os: ubuntu-20.04, r: 'devel'}
-          - {os: ubuntu-20.04, r: 'oldrel'}
-          - {os: ubuntu-20.04, r: '3.6'}
+          #
+          - {os: ubuntu-latest, r: 'release'}
+          - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
+          - {os: ubuntu-latest, r: 'oldrel'}
+          - {os: ubuntu-latest, r: '3.6'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
