+++ .github/workflows/check-full.yaml
@@ -26,12 +26,11 @@ jobs:
           - {os: windows-latest, r: 'devel'}
           - {os: windows-latest, r: 'release'}
           - {os: windows-latest, r: 'oldrel'}
-          #- {os: macOS-latest, r: 'devel'}
+          - {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
-          #- {os: macOS-latest, r: 'oldrel'}
-          #
+          - {os: macOS-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: 'release'}
-          - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
+          #- {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: '3.6'}
 
