+++ R/cffr-package.R
@@ -10,4 +10,5 @@ NULL
 # import stuffs
 #' @importFrom utils download.file as.person read.csv person
 #' installed.packages  packageDescription citation available.packages
+#' modifyList
 NULL
