+++ NAMESPACE
@@ -19,6 +19,7 @@ importFrom(utils,available.packages)
 importFrom(utils,citation)
 importFrom(utils,download.file)
 importFrom(utils,installed.packages)
+importFrom(utils,modifyList)
 importFrom(utils,packageDescription)
 importFrom(utils,person)
 importFrom(utils,read.csv)
