+++ .github/workflows/check-full.yaml
@@ -26,11 +26,11 @@ jobs:
           - {os: windows-latest, r: 'devel'}
           - {os: windows-latest, r: 'release'}
           - {os: windows-latest, r: 'oldrel'}
-          #- {os: macOS-latest, r: 'devel'}
+          - {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
-          #- {os: macOS-latest, r: 'oldrel'}
+          - {os: macOS-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: 'release'}
-          #- {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
+          - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: '3.6'}
 
@@ -47,7 +47,7 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
+          use-public-rspm: false
 
       - uses: r-lib/actions/setup-r-dependencies@master
         with:
