+++ .github/workflows/check-full.yaml
@@ -26,13 +26,16 @@ jobs:
           - {os: windows-latest, r: 'devel'}
           - {os: windows-latest, r: 'release'}
           - {os: windows-latest, r: 'oldrel'}
-          #- {os: macOS-latest, r: 'devel'}
+          - {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
-          #- {os: macOS-latest, r: 'oldrel'}
+          - {os: macOS-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: 'release'}
           - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest, r: 'oldrel'}
-          - {os: ubuntu-latest, r: '3.6'}
+          - {os: ubuntu-latest, r: 'oldrel-2'}
+          - {os: ubuntu-latest, r: 'oldrel-3'}
+          - {os: ubuntu-18.04, r: '3.6'}
+          - {os: ubuntu-18.04, r: '3.3'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
