+++ .github/workflows/check-full.yaml
@@ -28,14 +28,13 @@ jobs:
           - {os: windows-latest, r: 'oldrel'}
           - {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
-          - {os: macOS-latest, r: 'oldrel'}
+          #- {os: macOS-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: 'release'}
           - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest, r: 'oldrel'}
-          - {os: ubuntu-latest, r: 'oldrel-2'}
-          - {os: ubuntu-latest, r: 'oldrel-3'}
+          - {os: ubuntu-18.04, r: 'oldrel-2'}
           - {os: ubuntu-18.04, r: '3.6'}
-          - {os: ubuntu-18.04, r: '3.3'}
+
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
@@ -50,7 +49,7 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: false
+          use-public-rspm: true
 
       - uses: r-lib/actions/setup-r-dependencies@master
         with:
