### 2021-10-04 8972a6f
Small fix on citation parsing
