### 2021-10-06 4edbd6e
Improve logo
