+++ data-raw/logo.R
@@ -3,14 +3,26 @@ library(hexSticker)
 font_add_google("Roboto", "roboto")
 
 showtext_auto()
-sticker("data-raw/network.png",
+
+
+# Subplot
+img <- magick::image_read("data-raw/network.png")
+g <- grid::rasterGrob(img, interpolate = TRUE)
+
+p <- ggplot2::ggplot() +
+  ggplot2::annotation_custom(g, xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf) +
+  ggplot2::theme_void()
+
+
+p
+sticker(p,
   package = "cffr",
   p_family = "roboto",
   p_fontface = "bold",
-  s_width = .7,
-  s_height = .7,
+  s_width = 1.4,
+  s_height = 1.4,
   s_x = 0.9,
-  s_y = .75,
+  s_y = .77,
   p_color = "#1892E3",
   p_size = 30,
   p_y = 1.6,
