+++ .github/workflows/check-full.yaml
@@ -6,7 +6,9 @@
 # usethis::use_github_action("check-standard") will install it.
 on:
   push:
+    branches: [main, master]
   pull_request:
+    branches: [main, master]
   schedule:
       - cron:  '30 11 * * 1,4'
 
