+++ .github/workflows/check-full.yaml
@@ -28,7 +28,7 @@ jobs:
           - {os: windows-latest, r: 'devel'}
           - {os: windows-latest, r: 'release'}
           - {os: windows-latest, r: 'oldrel'}
-          - {os: macOS-latest, r: 'devel'}
+          #- {os: macOS-latest, r: 'devel'}
           - {os: macOS-latest, r: 'release'}
           #- {os: macOS-latest, r: 'oldrel'}
           - {os: ubuntu-latest, r: 'release'}
