### 2021-10-06 c1a1ba4
Improve logo (#9)
* Improve logo

* Move site to link on main
