+++ pkgdown/_pkgdown.yml
@@ -7,7 +7,8 @@ template:
       creator: "@dhernangomez"
       card: summary_large_image
 
-
+repo:
+  branch: main
 
 reference:
   - title: Main function
