### 2021-10-11 a87b0f4
Not run tests in parallel
It breaks the GHA check on all packages
