+++ dev/CITATION.cff
@@ -1,15 +1,0 @@
-cff-version: 1.2.0
-message: If you use this software, please cite it using these metadata.
-title: 'codemetar: Generate ''CodeMeta'' Metadata for R Packages'
-version: 0.1.6
-authors:
-- family-names: Ok
-  given-names: John
-- family-names: Doe
-  given-names: Jane
-- family-names: Doo
-  given-names: Jane
-abstract: The 'Codemeta' Project defines a 'JSON-LD' format for describing software
-  metadata, as detailed at <https://codemeta.github.io>. This package provides utilities
-  to generate, parse, and modify 'codemeta.json' files automatically for R packages,
-  as well as tools and examples for working with 'codemeta.json' 'JSON-LD' more generally.
