### 2021-10-13 fb04e6f
Update and fix some tests
