+++ man/cffr-package.Rd
@@ -9,10 +9,9 @@
 \if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
 
 The Citation File Format version 1.2.0 (Druskat, S. et al. (2021)
-\href{https://doi.org/10.5281/zenodo.5171937}{\url{doi:10.5281/zenodo.5171937}} is
-a human and machine readable file format which provides citation metadata
-for software. This package provides core utilities to generate and validate
-this metadata.
+\doi{10.5281/zenodo.5171937} is a human and machine readable file format
+which provides citation metadata for software. This package provides core
+utilities to generate and validate this metadata.
 }
 \seealso{
 Useful links:
