### 2021-10-21 a00c7ad
Update citations of the package
