+++ tests/testthat/test-cff_description.R
@@ -175,6 +175,10 @@ test_that("Search package on CRAN", {
 
 
 test_that("Search package on r-universe", {
+  
+  skip_on_cran()
+  skip_if_offline()
+  
   basic_path <- system.file("examples/DESCRIPTION_basic",
     package = "cffr"
   )
