+++ .github/workflows/check-full.yaml
@@ -34,8 +34,8 @@ jobs:
           - {os: ubuntu-latest, r: 'release'}
           - {os: ubuntu-latest, r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest, r: 'oldrel'}
-          - {os: ubuntu-18.04, r: 'oldrel-2'}
-          - {os: ubuntu-18.04, r: '3.6'}
+          - {os: ubuntu-latest, r: 'oldrel-2'}
+          - {os: ubuntu-latest, r: '3.6'}
 
 
     env:
