+++ R/cff.R
@@ -4,7 +4,7 @@
 #' information.
 #'
 #' @name cff
-#' @return 
+#' @return
 #' A `cff` object. Under the hood, a `cff` object is a regular [list()]
 #' object with a special [print()] method.
 #'
