+++ man/cff.Rd
@@ -18,7 +18,8 @@ as.cff(x)
 \item{x}{a character string for the \code{\link{as.cff}} default method}
 }
 \value{
-A \code{cff} list object.
+A \code{cff} object. Under the hood, a \code{cff} object is a regular \code{\link[=list]{list()}}
+object with a special \code{\link[=print]{print()}} method.
 }
 \description{
 A class and utility methods for reading, creating and holding CFF
