### 2021-10-28 ee6ebb7
Fix some citations
