+++ bibtex.bib
@@ -6,14 +6,14 @@
 	doi          = {10.5281/zenodo.5171937}
 }
 @misc{codemeta_2_0,
-	title        = {CodeMeta: an exchange schema for software metadata},
+	title        = {{CodeMeta}: an exchange schema for software metadata},
 	author       = {Matthew B. Jones and Carl Boettiger and Abby Cabunoc Mayes and Arfon Smith and Peter Slaughter and Kyle Niemeyer and Yolanda Gil and Martin Fenner and Krzysztof Nowak and Mark Hahnel and Luke Coy and Alice Allen and Mercè Crosas and Ashley Sands and Neil Chue Hong and Patricia Cruse and Daniel S. Katz and Carole Goble},
 	year         = 2017,
 	doi          = {10.5063/schema/codemeta-2.0},
 	version      = {2.0}
 }
 @misc{aligning_codemeta,
-	title        = {Aligning the CodeMeta vocabulary for scientific software with schema.org},
+	title        = {Aligning the {CodeMeta} vocabulary for scientific software with {schema.org}},
 	author       = {Martin Fenner},
 	year         = 2021,
 	month        = May,
