+++ README.Rmd
@@ -1,8 +1,5 @@
 ---
 output: github_document
-editor_options: 
-  markdown: 
-    wrap: 80
 bibliography: bibtex.bib
 ---
 
@@ -17,7 +14,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# cffr <a href='https://dieghernan.github.io/cffr/'><img src='man/figures/logo.png' align="right" height="139" /></a>
+# cffr <a href='https://dieghernan.github.io/cffr/'><img src="man/figures/logo.png" align="right" height="139"/></a>
 
 <!-- badges: start -->
 
@@ -35,26 +32,34 @@ bytes](https://img.shields.io/github/languages/code-size/dieghernan/cffr)
 
 <!-- badges: end -->
 
+## What is a `CITATION.cff` file?
+
 [Citation File Format (CFF)](https://citation-file-format.github.io/)
 [@Druskat_Citation_File_Format_2021] (v1.2.0) are plain text files with human-
 and machine-readable citation information for software (and datasets). Code
 developers can include them in their repositories to let others know how to
 correctly cite their software.
 
-When you put a `CITATION.cff` file in the default branch of your GitHub
-repository, it is automatically linked from the repository landing page, and the
-citation information is rendered on the repository page, and also provided as
-BibTeX snippet which users can simply copy.
+This format is becoming popular within the software citation ecosystem. Recently
+[GitHub](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files),
+[Zenodo](https://twitter.com/ZENODO_ORG/status/1420357001490706442) and
+[Zotero](https://twitter.com/zotero/status/1420515377390530560) have included
+full support of this citation format. GitHub support is of special interest:
 
-**cffr** provides utilities to generate, parse, modify and validate
-`CITATION.cff` files automatically for **R** packages, as well as tools and
-examples for working with .cff more generally.
+*See [Customize your repository/About CITATION
+files](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files)*
 
-**cffr** maximizes the data extraction by using both the `DESCRIPTION` file and
-the `CITATION` file (if present) of your package. Note that **cffr** works best
-if your package pass `R CMD check/devtools::check()`.
 
-## Related projects
+> When you add a CITATION.cff file to the default branch of your repository, it is
+> automatically linked from the repository landing page. This makes it easy for
+> other users to cite your software project, using the information you've
+> provided.
+
+```{r echo=FALSE, out.width="80%", fig.align='center', fig.alt="citation-link"}
+knitr::include_graphics("vignettes/citation-link.png")
+```
+
+### Related projects
 
 [The CodeMeta Project](https://codemeta.github.io/) [@codemeta_2_0] creates a
 concept vocabulary that can be used to standardize the exchange of software
@@ -66,7 +71,17 @@ year, and venue [@aligning_codemeta]. The packages
 [**codemetar**](https://github.com/ropensci/codemetar) allows to generate
 `codemeta.json` files from R packages metadata.
 
-## Installation
+## The cffr package
+
+**cffr** provides utilities to generate, parse, modify and validate
+`CITATION.cff` files automatically for **R** packages, as well as tools and
+examples for working with .cff more generally.
+
+**cffr** maximizes the data extraction by using both the `DESCRIPTION` file and
+the `CITATION` file (if present) of your package. Note that **cffr** works best
+if your package pass `R CMD check/devtools::check()`.
+
+### Installation
 
 You can install the developing version of **cffr** with:
 
@@ -86,7 +101,7 @@ options(repos = c(
[…]
