### 2021-10-30 #463 @emilyriederer
@dieghernan , I know I shared a *lot* in the message above, so I wanted to highlight one more specific item for your consideration. I think this package could be of very high interest to the rOpenSci community since it would apply generally across all domains. If you'd be interested in thinking about how this package could be highlighted in an rOpenSci blog post or [tech note](https://ropensci.org/technotes/), it might be worth having a discussion with @ropensci/blog-editors

I think it could be great to highlight how this package is similar in spirit to well-known tools like `devtools` and `usethis` and smartly reuses repository context to create a general-purpose citation file. This could also help raise awareness for GitHub's citation feature and the benefits that you see in it since it is relatively new.
