### 2021-11-01 #463 @emilyriederer
Hi @dieghernan - my apologies. You should have now received an invitation to join a team within the `ropensci` organization. Please accept that and try the transfer. After it is complete, I will go in and make you an admin on the transferred repo.
