### 2021-11-01 #463 @stefaniebutland
Hi @dieghernan, I'm rOpenSci's Community Manager. Congratulations on cffr passing review! Yes a post noted by Emily would be really valuable. Narrative and technical guidelines are in https://blogguide.ropensci.org/. If you're interested, please suggest when you would like to submit a draft. I recognize this is extra work, so no pressure intended here :-)

My colleague @steffilazerte would review it and suggest a publication date.
