### 2021-11-02 #463 @dieghernan
Hi @stefaniebutland, @steffilazerte

I’ll have a look and let you know a feasible date for the blog post.

I was also thinking on submitting the package to CRAN, and I wonder if submitting it before or after you publish the post. The package is almost ready for it.

What are your views on this? Regards
