### 2021-11-02 #463 @stefaniebutland
I can't think of a specific reason to choose publishing before or after CRAN (Steffi how about you?). It's up to you. We could have your post in a pull request, reviewed and ready to publish once on CRAN if you prefer that.
