### 2021-11-02 #463 @steffilazerte
I think it's up to you as well. I can't imagine that CRAN would have you change anything major that would affect the user experience, and we generally have you link to the ropensci docs page (not to the CRAN page), so it shouldn't matter in the blog post.
