### 2021-11-04 #463 @dieghernan
Hi @steffilazerte @emilyriederer

I think I could send a draft post by the end of the next week. So assuming at least one week for reviewing and that the posts are usually published on Tuesday (as per the [Blogguide](https://blogguide.ropensci.org/content.html) the first feasible publishing date would be **Nov. 23 2021**.

Let me know what you think,

Regads
