### 2021-11-04 #463 @stefaniebutland
Just fyi, @maelle & others wrote this post in Feb 2021: [Make Your R Package Easier to Cite](https://ropensci.org/blog/2021/02/16/package-citation/). So you know what rOpenSci has said on the topic before.

Steffi and I agreed Nov 23 is a good tentative publication date.
