### 2021-11-05 #463 @dieghernan
And it is on CRAN! https://cran.r-project.org/package=cffr
